% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{check.ca.plot}
\alias{check.ca.plot}
\title{check.ca.plot}
\usage{
check.ca.plot(data, x, y)
}
\arguments{
\item{x,y:}{The numbers of the dimensions to be plotted (from 1 to 5).}

\item{data:}{Dataframe containing the contingency table to be analysed.}
}
\description{
This function returns the Correspondence Analysis scatterplot of both row and column categories.
}
\examples{
check.ca.plot (mydata,1,2): plot the Correspondence Analysis scatterplot of the 'mydata' dataframe, displaying the 1 and 2 dimension.
}
\keyword{analysis,}
\keyword{correspondence}
\keyword{plot}

