% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{evaluate}
\alias{evaluate}
\title{evaluate}
\usage{
evaluate(data, x = 1, y = 2, which)
}
\arguments{
\item{x, y:}{The dimensions to be plotted (1st and 2nd by default).}

\item{data:}{Dataframe containing the contingency table to be analysed.}

\item{which:}{use R for rows, C for colums.}
}
\description{
This function returns the Correspondence Analysis scatterplot of row/column categories with a 2nd order polynomial fit. The Rsquared value is reported in the plot's subtitle. 
The function allows gauging to what extent the seriation structure (if any) embedded in the data approximates to a 'perfect' seriation.
}
\examples{
data("cemetery"): load the dataset 'cemetery'
evaluate(cemetery,1,2,which='R'): plot the Correspondence Analysis scatterplot of row categories (using the 'cemetery' dataset), displaying the 1 and 2 dimensions; a 2nd order polynomial fit is displayed and the Rsquared value is reported.

}
\keyword{2nd}
\keyword{evaluation,}
\keyword{fit}
\keyword{order}
\keyword{polynomial,}
\keyword{seriation,}

