% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{plot.clusters.cols}
\alias{plot.clusters.cols}
\title{plot.clusters.cols}
\usage{
\method{plot}{clusters.cols}(data, x, y)
}
\arguments{
\item{x,y:}{The numbers of the dimensions to be plotted (from 1 to 5).}

\item{data:}{Dataframe containing the contingency table to be analysed.}
}
\description{
This function returns the Correspondence Analysis scatterplot of column categories, with points' colour according to the clusters to which points belong.
An Excel file is also created (in the R working directory) containing the cluster membership for each colum category.
}
\examples{
plot.clusters.cols (mydata,1,2): plot the Correspondence Analysis scatterplot of column categories (using the 'mydata' dataframe), displaying the 1 and 2 dimension.
}
\keyword{analysis}
\keyword{cluster,}
\keyword{columns,}
\keyword{correspondence}

