% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{plot.clusters.rows}
\alias{plot.clusters.rows}
\title{plot.clusters.rows}
\usage{
\method{plot}{clusters.rows}(data, x, y)
}
\arguments{
\item{x,y:}{The numbers of the dimensions to be plotted (from 1 to 5).}

\item{data:}{Dataframe containing the contingency table to be analysed.}
}
\description{
This function returns the Correspondence Analysis scatterplot of row categories, with points' colour according to the clusters to which points belong.
An Excel file is also created (in the R working directory) containing the cluster membership for each row category.
}
\examples{
plot.clusters.rows (mydata,1,2): plot the Correspondence Analysis scatterplot of row categories (using the 'mydata' dataframe), displaying the 1 and 2 dimension.
}
\keyword{analysis}
\keyword{clusters,}
\keyword{correspondence}
\keyword{plot,}
\keyword{rows,}

