% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{sort.table}
\alias{sort.table}
\title{sort.table}
\usage{
sort.table(data, dim)
}
\arguments{
\item{data:}{Dataframe containing the contingency table to be analysed.}

\item{dim:}{The number of the dimension (from 1 to 5) to be used to sort the contingency table (i.e., to perfom the seriation).}
}
\description{
This function allows to sort the input contingency table according to the score of row and column categories on the selected Correspondence Analysis dimension.
Two plots are returned, displaying the sorted contingency table. Note that the two plots are the same, except for the fact that one is the transposed version of the other.
An Excel file (named CAseriation.xlsx) is also created in the R working directory, containing the original dataset, the sorted table, and the sorted table converted into presence/absence data.
}
\examples{
sort.table (mydata,1): sort the contingency table (contained in the dataframe 'mydata') according to the score of row and column categories on the 1 dimension of Correspondence Analysis.
}
\keyword{analysis}
\keyword{contingency}
\keyword{correspondence}
\keyword{scores,}
\keyword{sort,}
\keyword{table,}

