% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_ca_plot.R
\name{check.ca.plot}
\alias{check.ca.plot}
\title{check.ca.plot}
\usage{
check.ca.plot(data, x = 1, y = 2)
}
\arguments{
\item{x, y:}{The dimensions to be plotted (1st and 2nd by default).}

\item{data:}{Dataframe containing the contingency table to be analysed.}
}
\description{
This function returns the Correspondence Analysis scatterplot of both row and column categories.
}
\examples{
data("perfect_seriation"): load the dataset 'perfect_seriation'
check.ca.plot (perfect_seriation,1,3): plot the Correspondence Analysis scatterplot of the 'perfect_seriation' dataframe, displaying the 1 and 3 dimensions.

}
\keyword{analysis,}
\keyword{correspondence}
\keyword{plot}

