% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_clusters_rows.R
\name{plot.clusters.rows}
\alias{plot.clusters.rows}
\title{plot.clusters.rows}
\usage{
\method{plot}{clusters.rows}(data, x = 1, y = 2, bp = FALSE, varwidth = FALSE)
}
\arguments{
\item{x, y:}{The dimensions to be plotted (1st and 2nd by default).}

\item{data:}{Dataframe containing the contingency table to be analysed.}

\item{bp:}{logical value (TRUE/FALSE). TRUE displays boxplots of row coordinates by cluster, in order to visually gauge the degree of separation between clusters. FALSE is default.}

\item{varwidth:}{logical value (TRUE/FALSE). TRUE sets the width of the boxplots proportional to the groups size. This helps in eyeballing the size of the clusters being compared. FALSE is default.}
}
\description{
This function returns the Correspondence Analysis scatterplot of row categories, with points' colour according to the clusters to which points belong.
A tab delimited .txt file is also created (in the R working directory) containing the cluster membership for each row category.
}
\examples{
data("perfect_seriation"): load the dataset 'perfect_seriation'
plot.clusters.rows (perfect_seriation,1,3): plot the Correspondence Analysis scatterplot of row categories (using the 'perfect_seriation' dataframe), displaying the 1 and 3 dimensions.
plot.clusters.rows (cemetery,1,2,bp=TRUE,varwidth=TRUE): plot the Correspondence Analysis scatterplot of row categories (using the 'cemetery' dataframe), displaying the 1 and 2 dimension. Boxplots of row coordinates (on the selected dimensions) by cluster are also displayed, with size proportional to the size of the clusters.
}
\keyword{analysis}
\keyword{cluster,}
\keyword{columns,}
\keyword{correspondence}

